+++ pkg/utils/kubernetes/owners.go
@@ -1,6 +1,8 @@
 package kubernetes
 
 import (
+	"fmt"
+
 	"github.com/samber/lo"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -65,20 +67,39 @@ func SetOwnerForObjectThroughLabels[managingObject managingObjectT](obj client.O
 // GetManagedByLabelSet returns a map of labels with the provided object's metadata.
 // These can be applied to other objects that are owned by the object provided as an argument.
 func GetManagedByLabelSet[managingObject managingObjectT](object managingObject) map[string]string {
-	var kindLabel string
-	switch any(object).(type) {
-	case *gatewayv1.Gateway:
-		kindLabel = consts.GatewayManagedLabelValue
-	case *operatorv1beta1.ControlPlane:
-		kindLabel = consts.ControlPlaneManagedLabelValue
-	case *operatorv1beta1.DataPlane:
-		kindLabel = consts.DataPlaneManagedLabelValue
-	}
-
 	return map[string]string{
-		consts.GatewayOperatorManagedByLabel:          kindLabel,
+		consts.GatewayOperatorManagedByLabel:          getKindLabel(object),
 		consts.GatewayOperatorManagedByNamespaceLabel: object.GetNamespace(),
 		consts.GatewayOperatorManagedByNameLabel:      object.GetName(),
+	}
+}
+
+// GetLegacyManagedByLabel returns a map of legacy labels with the provided object's metadata.
+// These can be applied to other objects that are owned by the object provided as an argument.
+func GetLegacyManagedByLabel[managingObject managingObjectT](object managingObject) map[string]string {
+	return map[string]string{
+		consts.GatewayOperatorManagedByLabel: getKindLabel(object),
+	}
+}
+
+// GetLegacyManagedByLabelSet returns a map of legacy labels with the provided object's metadata.
+// These can be applied to other objects that are owned by the object provided as an argument.
+func GetLegacyManagedByLabelSet[managingObject managingObjectT](object managingObject) map[string]string {
+	m := GetLegacyManagedByLabel(object)
+	m["app"] = object.GetName()
+	return m
+}
+
+func getKindLabel[managingObject managingObjectT](object managingObject) string {
+	switch any(object).(type) {
+	case *gatewayv1.Gateway:
+		return consts.GatewayManagedLabelValue
+	case *operatorv1beta1.ControlPlane:
+		return consts.ControlPlaneManagedLabelValue
+	case *operatorv1beta1.DataPlane:
+		return consts.DataPlaneManagedLabelValue
+	default:
+		return fmt.Sprintf("unknown-kind-%s", object.GetObjectKind().GroupVersionKind().Kind)
 	}
 }